+++ src/ui.js
@@ -14,6 +14,9 @@ var ui = {};
 (function ($, ui) {
     "use strict";
 
+    var barClasses = ["danger", "warning", "success"],
+        statusClasses = ["error", "warning", "success"];
+
     ui.getContainer = function (options, $el) {
         var $container;
 
@@ -139,7 +142,7 @@ var ui = {};
         $.each(ui.possibleProgressBarClasses, function (idx, value) {
             $bar.removeClass(cssPrefix + "bar-" + value);
         });
-        $bar.addClass(cssPrefix + "bar-" + cssClass);
+        $bar.addClass(cssPrefix + "bar-" + barClasses[cssClass]);
         $bar.css("width", percentage + '%');
     };
 
@@ -157,6 +160,20 @@ var ui = {};
         $errors.html(html);
     };
 
+    ui.updateFieldStatus = function (options, $el, cssClass) {
+        var targetClass = options.ui.bootstrap2 ? ".control-group" : ".form-group",
+            $container = $el.parents(targetClass).first();
+
+        $.each(statusClasses, function (idx, css) {
+            if (!options.ui.bootstrap2) { css = "has-" + css; }
+            $container.removeClass(css);
+        });
+
+        cssClass = statusClasses[cssClass];
+        if (!options.ui.bootstrap2) { cssClass = "has-" + cssClass; }
+        $container.addClass(cssClass);
+    };
+
     ui.percentage = function (score, maximun) {
         var result = Math.floor(100 * score / maximun);
         result = result < 0 ? 0 : result;
@@ -165,30 +182,33 @@ var ui = {};
     };
 
     ui.updateUI = function (options, $el, score) {
-        var barCss, barPercentage, verdictText;
+        var cssClass, barPercentage, verdictText;
 
         barPercentage = ui.percentage(score, options.ui.scores[3]);
         if (score <= 0) {
-            barCss = "danger";
+            cssClass = 0;
             verdictText = "";
         } else if (score < options.ui.scores[0]) {
-            barCss = "danger";
+            cssClass = 0;
             verdictText = options.ui.verdicts[0];
         } else if (score < options.ui.scores[1]) {
-            barCss = "danger";
+            cssClass = 0;
             verdictText = options.ui.verdicts[1];
         } else if (score < options.ui.scores[2]) {
-            barCss = "warning";
+            cssClass = 1;
             verdictText = options.ui.verdicts[2];
         } else if (score < options.ui.scores[3]) {
-            barCss = "warning";
+            cssClass = 1;
             verdictText = options.ui.verdicts[3];
         } else {
-            barCss = "success";
+            cssClass = 2;
             verdictText = options.ui.verdicts[4];
         }
 
-        ui.updateProgressBar(options, $el, barCss, barPercentage);
+        ui.updateProgressBar(options, $el, cssClass, barPercentage);
+        if (options.ui.showStatus) {
+            ui.updateFieldStatus(options, $el, cssClass);
+        }
         if (options.ui.showPopover) {
             // Popover can't be updated, it has to be recreated
             ui.initPopover(options, $el, verdictText);