+++ src/options.js
@@ -58,6 +58,7 @@ defaultOptions.rules.raisePower = 1.4;
 
 defaultOptions.ui = {};
 defaultOptions.ui.bootstrap2 = false;
+defaultOptions.ui.showProgressBar = true;
 defaultOptions.ui.showPopover = false;
 defaultOptions.ui.showStatus = false;
 defaultOptions.ui.spanError = function (options, key) {