+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Support zxcvbn for password scoring.
 - Support showing the password strength as a validation status in the password
   field.
+- Support hiding the progress bar, making it optional.
 
 ## 1.0.2